% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydRun_functions.R
\name{compute.hydro.ITC}
\alias{compute.hydro.ITC}
\title{Event hydrograph timing characteristics}
\usage{
compute.hydro.ITC(RunoffEvent)
}
\arguments{
\item{RunoffEvent}{A data frame containing a runoff event (format is 2 columns: datetime and stormflow).}
}
\value{
A list containing the time of runoff event \code{start}, \code{end}, \code{peak}, and \code{centroid}.

The unit of time characteristics is in hours.
}
\description{
\code{compute.hydro.ITC} is used to calculate timing characteristics of a runoff event storm hydrograph.
}
\note{
This is an R implementation of a function from the MATLAB toolbox HydRun.
}
\examples{
events <- compute.hydro.ITC(RunoffEvent)
}
\references{
Tang, W., & Carey, S. K. (2017). HydRun: a MATLAB toolbox for rainfall runoff analysis. Hydrological Processes, 31(15), 2670-2682.
}
