% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydRun_functions.R
\name{compute.hyeto.ITC}
\alias{compute.hyeto.ITC}
\title{Event hyetograph timing characteristics}
\usage{
compute.hyeto.ITC(PrecipEvent)
}
\arguments{
\item{PrecipEvent}{A data frame containing a precipitation event  event (format is 2 columns: datetime and precipitation).}
}
\value{
A list containing the time of runoff event \code{start}, \code{end}, and \code{centroid}.

The unit of time characteristics is in hours.
}
\description{
\code{compute.hyeto.ITC} is used to calculate timing characteristics of a precipitation event hyetograph.
}
\note{
This is an R implementation of a function from the MATLAB toolbox HydRun.
}
\examples{
events <- compute.hyeto.ITC(PrecipEvent)
}
\references{
Tang, W., & Carey, S. K. (2017). HydRun: a MATLAB toolbox for rainfall runoff analysis. Hydrological Processes, 31(15), 2670-2682.
}
