% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydRun_functions.R
\name{compute.RR}
\alias{compute.RR}
\title{Event runoff ratio}
\usage{
compute.RR(PrecipEvent, RunoffEvent, drainageArea)
}
\arguments{
\item{PrecipEvent}{A data frame containing a precipitation event  event (format is 2 columns: datetime and precipitation).}

\item{RunoffEvent}{A data frame containing a runoff event  event (format is 2 columns: datetime and stormflow).}

\item{drainageArea}{The drainage area of the catchment.}
}
\value{
A list containing the event runoff ratio \code{runoffRatio}, precipitation volume \code{precip_vol}, 
and runoff volume \code{runoff_vol}.

The unit of the returned volumes is in m^3
}
\description{
\code{compute.RR} is used to calculate the runoff ratio of a rainfall-runoff event
}
\note{
This is an R implementation of a function from the MATLAB toolbox HydRun.

It is assumed that the input precipitation is measured in mm, runoff in m^3/s, and drainage area in km^2
}
\examples{
events <- compute.hydro.ITC(RunoffEvent)
}
\references{
Tang, W., & Carey, S. K. (2017). HydRun: a MATLAB toolbox for rainfall runoff analysis. Hydrological Processes, 31(15), 2670-2682.
}
