% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydRun_functions.R
\name{compute.TC}
\alias{compute.TC}
\title{Rainfall-runoff event timing characteristics}
\usage{
compute.TC(PrecipEvent, RunoffEvent)
}
\arguments{
\item{PrecipEvent}{A data frame containing a precipitation event  event (format is 2 columns: datetime and precipitation).}

\item{RunoffEvent}{A data frame containing a runoff event  event (format is 2 columns: datetime and stormflow).}
}
\value{
A list containing the rainfall duration \code{Tw}, response lag \code{TLR}, 
time of rise \code{Tr}, lag-to-peak \code{TLP}, centroid lag-to-peak \code{TLPC}, centroid lag \code{TLC},
time base \code{Tb}.

The unit of time characteristics is in hours.
}
\description{
\code{compute.TC} is used to calculate timing characteristics of a rainfall-runoff event.
}
\note{
This is an R implementation of a function from the MATLAB toolbox HydRun.
}
\examples{
events <- compute.TC(PrecipEvent, RunoffEvent)
}
\references{
Tang, W., & Carey, S. K. (2017). HydRun: a MATLAB toolbox for rainfall runoff analysis. Hydrological Processes, 31(15), 2670-2682.
}
