% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydRun_functions.R
\name{extract.precip}
\alias{extract.precip}
\title{Extract precipitation events}
\usage{
extract.precip(precip, MAG, Thr)
}
\arguments{
\item{precip}{A precipitation data frame (format is 2 columns: datetime and precipitation).}

\item{MAG}{The maximum allowable gap for an intermittent rainfall, expressed in hours.}

\item{Thr}{The minimum amount of precipitation.}
}
\value{
A list containing the precipitation events, \code{PrecipEvents} the number of precipitation events, \code{nPrecipEvent}, and the portion of NAs in a rainfall event, \code{pNA}.
}
\description{
Extracts precipitation events from a precipitation timeseries.
}
\note{
This is an R implementation of a function from the MATLAB toolbox HydRun.
}
\examples{
precipitation_events <- extract.precip(precip, 24, 0)
}
\references{
Tang, W., & Carey, S. K. (2017). HydRun: a MATLAB toolbox for rainfall runoff analysis. Hydrological Processes, 31(15), 2670-2682.
}
