% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydRun_functions.R
\name{extract.runoff}
\alias{extract.runoff}
\title{Extract runoff events}
\usage{
extract.runoff(
  stormflow,
  min_diff,
  return_ratio,
  b_slope,
  e_slope,
  sc,
  min_dur,
  dyslp
)
}
\arguments{
\item{stormflow}{A stream stormflow data frame (format is 2 columns: datetime and stormflow).}

\item{min_diff}{The minimum difference between the start (or end) of runoff and the runoff peak.}

\item{return_ratio}{Determines where runoff terminates: runoff is terminated when it declines below a dynamic threshold (peak discharge)*\code{return_ratio}.}

\item{b_slope}{The beginning slope (slope threshold used to cut flat head).}

\item{e_slope}{The ending slope (slope threshold to end event).}

\item{sc}{The smoothing coefficient, which determines the number of passes applied to \code{stormflow}.}

\item{min_dur}{The minimum duration of a runoff event, expressed as the number of timesteps.}

\item{dyslp}{The dynamic slope threshold used to cut the flat head and end the runoff event.}
}
\value{
A list containing the runoff events, \code{RunoffEvents}, and the number of runoff events, \code{nRunoffEvent}.
}
\description{
Extracts runoff events from a stormflow timeseries.
}
\note{
This is an R implementation of a function from the MATLAB toolbox HydRun.
}
\examples{
runoff_events <- extract.runoff(stormflow, 1.1, 0.001, 0.001, 0.35, 1, 4, 0.001)
}
\references{
Tang, W., & Carey, S. K. (2017). HydRun: a MATLAB toolbox for rainfall runoff analysis. Hydrological Processes, 31(15), 2670-2682.
}
