% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydRun_functions.R
\name{find.tp}
\alias{find.tp}
\title{Identify turning points in a line}
\usage{
find.tp(X)
}
\arguments{
\item{X}{The line being assessed for turning points.}
}
\value{
A data frame (format is 2 columns: index of turning points in \code{X} and labels for peaks (1) and valleys (0)).
}
\description{
Identifies turning/inflection points in a line.
}
\note{
This is an R implementation of a function from the MATLAB toolbox HydRun.
}
\examples{
turning_points <- find.tp(response)
}
\references{
Tang, W., & Carey, S. K. (2017). HydRun: a MATLAB toolbox for rainfall runoff analysis. Hydrological Processes, 31(15), 2670-2682.
}
