% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydRun_functions.R
\name{get.mode}
\alias{get.mode}
\title{Calculate mode of vector}
\usage{
get.mode(v)
}
\arguments{
\item{v}{A vector to calculate the mode of.}
}
\value{
The mode
}
\description{
Calculate mode of vector.
}
\examples{
vector_mode <- get.mode(V)
}
