% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydRun_functions.R
\name{separate.baseflow}
\alias{separate.baseflow}
\title{Baseflow separation}
\usage{
separate.baseflow(hydrograph, filter, passes)
}
\arguments{
\item{hydrograph}{A stream hydrograph data frame (format is 2 columns: datetime and discharge).}

\item{filter}{The filter coefficient used in baseflow separation - range 0.9 - 0.95.}

\item{passes}{The number of filter passes.}
}
\value{
A list containing stormflow and baseflow of \code{hydrograph}.
}
\description{
\code{separate.baseflow} is used to separate a stream hydrograph into baseflow and stormflow components.
}
\note{
This is an R implementation of a function from the MATLAB toolbox HydRun.
}
\examples{
hydrograph_separated <- separate.baseflow(hydrograph, 0.9, 3)
}
\references{
Tang, W., & Carey, S. K. (2017). HydRun: a MATLAB toolbox for rainfall runoff analysis. Hydrological Processes, 31(15), 2670-2682.

Fuka, D. R., Walter, M. T., Archibald, J. A., Steenhuis, T. S., Easton, Z. M., Fuka, M. D., & KeepSource, T. R. U. E. (2014). Package EcoHydRology.
}
