% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydRun_functions.R
\name{smooth.curve}
\alias{smooth.curve}
\title{Smooth runoff response data}
\usage{
smooth.curve(X, Pass)
}
\arguments{
\item{X}{The response data vector.}

\item{Pass}{The number of filter passes used on \code{X}.}
}
\value{
A smoothed response vector.
}
\description{
Smooths runoff response data for runoff event delineation using a moving average filter (window is 3).
}
\note{
This is an R implementation of a function from the MATLAB toolbox HydRun.
}
\examples{
smoothed_response <- smooth.curve(response, 3)
}
\references{
Tang, W., & Carey, S. K. (2017). HydRun: a MATLAB toolbox for rainfall runoff analysis. Hydrological Processes, 31(15), 2670-2682.
}
